Fixed build : moved indexer-dependent file from editor.
