+++ indexer/osm_editor.hpp
@@ -6,10 +6,10 @@
 #include "indexer/feature.hpp"
 #include "indexer/feature_meta.hpp"
 #include "indexer/mwm_set.hpp"
+#include "indexer/new_feature_categories.hpp"
 
 #include "editor/editor_config.hpp"
 #include "editor/editor_notes.hpp"
-#include "editor/new_feature_categories.hpp"
 #include "editor/xml_feature.hpp"
 
 #include "base/timer.hpp"
