+++ qt/create_feature_dialog.cpp
@@ -1,6 +1,6 @@
 #include "qt/create_feature_dialog.hpp"
 
-#include "editor/new_feature_categories.hpp"
+#include "indexer/new_feature_categories.hpp"
 
 #include <QtWidgets/QDialogButtonBox>
 #include <QtWidgets/QListWidget>
